# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.LoginRequest:
    properties:
      identifier:
        example: admin
        type: string
      password:
        example: admin123
        type: string
    required:
    - identifier
    - password
    type: object
  request_structs.AdminPermissionsDTO:
    properties:
      can_create_admin:
        description: Gestión del Staff (Escalamiento y Delegación)
        example: false
        type: boolean
      can_create_psi:
        description: Gestión de Colegiados (CRUD de Psicólogos)
        example: false
        type: boolean
      can_create_tags:
        description: Taxonomías (Clasificación de contenido)
        example: false
        type: boolean
      can_delete_admin:
        example: false
        type: boolean
      can_delete_psi:
        example: false
        type: boolean
      can_delete_publish:
        example: false
        type: boolean
      can_delete_tags:
        example: false
        type: boolean
      can_edit_tags:
        example: false
        type: boolean
      can_manage_notifications:
        example: false
        type: boolean
      can_publish:
        description: Sistema de Gestión de Contenidos (CMS)
        example: true
        type: boolean
      can_read_notifications:
        example: true
        type: boolean
      can_send_notifications:
        description: Motor de Mensajería y Alertas
        example: true
        type: boolean
      can_update_admin:
        example: false
        type: boolean
      can_update_psi:
        example: false
        type: boolean
      can_update_publish:
        example: true
        type: boolean
    type: object
  request_structs.CreateAdminRequest:
    properties:
      email:
        example: staff@colpsicarabobo.com
        type: string
      password:
        example: Segura123!
        minLength: 8
        type: string
      permissions:
        $ref: '#/definitions/request_structs.AdminPermissionsDTO'
      username:
        example: staff_admin
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  request_structs.UpdateAdminRequest:
    properties:
      email:
        example: nuevo@correo.com
        type: string
      id:
        description: |-
          ID es el identificador primario del administrador a afectar.
          Nota Arquitectónica: En controladores robustos, este ID en el body debe
          ser verificado cruzándolo con el ID de la URL (/api/admin/:id) para
          evitar ataques de manipulación de Payload.
        type: string
      is_active:
        example: true
        type: boolean
      password:
        minLength: 8
        type: string
      permissions:
        allOf:
        - $ref: '#/definitions/request_structs.AdminPermissionsDTO'
        description: Sub-estructura que será evaluada por el Motor de Permisos.
      username:
        example: nuevo_username
        type: string
    required:
    - id
    type: object
...
  service.BirthdayInfoProjection:
    properties:
      day:
        type: integer
      first_name:
        type: string
      fpv:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      month:
        description: Mes y día del cumpleaños (para mostrarlos como fecha sin exponer
          el año real).
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend para la gestión del Colegio de Psicólogos de Carabobo.
  title: ColPsiCarabobo API
  version: "1.0"
paths:
  /admin/create:
    post:
      consumes:
      - application/json
      description: Registra un nuevo miembro del staff administrativo verificando
        la jerarquía de permisos.
      parameters:
      - description: Datos del nuevo administrador y permisos
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_structs.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: creado correctamente'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: datos inválidos'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: violación de jerarquía'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Crear un nuevo administrador
      tags:
      - Administración - Gestión
  /admin/delete/{id}:
    delete:
      description: Realiza un borrado lógico del administrador. No permite auto-eliminación
        ni borrar SUDO.
      parameters:
      - description: UUID del administrador
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: eliminado correctamente'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: ID inválido'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: permisos insuficientes'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar administrador (Soft-delete)
      tags:
      - Administración - Gestión
  /admin/list:
    get:
      description: Retorna una lista paginada de administradores con soporte para
        búsqueda y caché.
      parameters:
      - description: 'Número de página (def: 1)'
        in: query
        name: page
        type: integer
      - description: 'Registros por página (def: 10)'
        in: query
        name: limit
        type: integer
      - description: Búsqueda por username o email
        in: query
        name: search
        type: string
      - description: Filtrar por estado activo (true/false)
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: data, total, page, limit
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: fallo interno'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar administradores
      tags:
      - Administración - Gestión
...
  /admin/psi/birthdays:
    get:
      description: Retorna los agremiados que cumplen años hoy o en los próximos días
        y han autorizado el aviso de cumpleaños (opt-in).
      parameters:
      - description: 'rango: ''today'' (hoy) o ''week'' (próximos 7 días). Default:
          today'
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.BirthdayInfoProjection'
            type: array
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Aviso de cumpleaños (Admin)
      tags:
      - Administración - Psicólogos
...
  /admin/update:
    patch:
      consumes:
      - application/json
      description: Modifica los datos y permisos de un administrador. Actualiza automáticamente
        la auditoría.
      parameters:
      - description: Campos parciales a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_structs.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: actualizado correctamente'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: permiso denegado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar administrador
      tags:
      - Administración - Gestión
  /admin/validate:
    get:
      description: Devuelve 200 si el token JWT es válido y la sesión sigue activa;
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Validar sesión de administrador
      tags:
      - Administración - Sesión
  /auth/login:
    post:
      consumes:
      - application/json
      description: Valida las credenciales del administrador (email o username) y
        retorna un JWT dinámico.
      parameters:
      - description: Credenciales de acceso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message, token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: formato inválido'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: credenciales inválidas'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Iniciar sesión administrativo
      tags:
      - Administración - Auth
  /auth/logout:
    post:
      description: Invalida la sesión del administrador a nivel de servidor (Stateful
        Logout).
      responses:
        "200":
          description: 'message: sesión cerrada'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: no autenticado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cerrar sesión administrativo
      tags:
      - Administración - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"